objectstate: rte: extract helpers
move helpers in per-mode files.
This is meant to keep the filesize at bay
and to enable future code extraction.

No intended change in behavior.

(cherry picked from commit 6668e12731dfe34ad71588e80c7d828d73ab7cdf)
