+++ pkg/objectstate/rte/nodegroup.go
@@ -1,0 +1,94 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Copyright 2025 Red Hat, Inc.
+ */
+
+package rte
+
+import (
+	"context"
+	"fmt"
+
+	appsv1 "k8s.io/api/apps/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	nropv1 "github.com/openshift-kni/numaresources-operator/api/numaresourcesoperator/v1"
+	nodegroupv1 "github.com/openshift-kni/numaresources-operator/api/numaresourcesoperator/v1/helper/nodegroup"
+	"github.com/openshift-kni/numaresources-operator/pkg/objectnames"
+	"github.com/openshift-kni/numaresources-operator/pkg/objectstate"
+	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/compare"
+	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/merge"
+)
+
+func updateFromClientTreeNodeGroup(ret *ExistingManifests, ctx context.Context, cli client.Client, instance *nropv1.NUMAResourcesOperator, tree nodegroupv1.Tree, namespace string) {
+	generatedName := objectnames.GetComponentName(instance.Name, *tree.NodeGroup.PoolName)
+	key := client.ObjectKey{
+		Name:      generatedName,
+		Namespace: namespace,
+	}
+	ds := &appsv1.DaemonSet{}
+	dsm := daemonSetManifest{}
+	if dsm.daemonSetError = cli.Get(ctx, key, ds); dsm.daemonSetError == nil {
+		dsm.daemonSet = ds
+	}
+	ret.daemonSets[generatedName] = dsm
+}
+
+func stateFromNodeGroup(em *ExistingManifests, mf Manifests, tree nodegroupv1.Tree) []objectstate.ObjectState {
+	var ret []objectstate.ObjectState
+	var existingDs client.Object
+	var loadError error
+
+	poolName := *tree.NodeGroup.PoolName
+
+	generatedName := objectnames.GetComponentName(em.instance.Name, poolName)
+	existingDaemonSet, ok := em.daemonSets[generatedName]
+	if ok {
+		existingDs = existingDaemonSet.daemonSet
+		loadError = existingDaemonSet.daemonSetError
+	} else {
+		loadError = fmt.Errorf("failed to find daemon set %s/%s", mf.Core.DaemonSet.Namespace, mf.Core.DaemonSet.Name)
+	}
+
+	desiredDaemonSet := mf.Core.DaemonSet.DeepCopy()
+	desiredDaemonSet.Name = generatedName
+
+	var updateError error
+	desiredDaemonSet.Spec.Template.Spec.NodeSelector = map[string]string{
+		HyperShiftNodePoolLabel: poolName,
+	}
+
+	gdm := GeneratedDesiredManifest{
+		ClusterPlatform:       em.plat,
+		MachineConfigPool:     nil,
+		NodeGroup:             tree.NodeGroup.DeepCopy(),
+		DaemonSet:             desiredDaemonSet,
+		IsCustomPolicyEnabled: em.customPolicyEnabled,
+	}
+
+	err := em.updater(poolName, &gdm)
+	if err != nil {
+		updateError = fmt.Errorf("daemonset for pool %q: update failed: %w", poolName, err)
+	}
+
+	ret = append(ret, objectstate.ObjectState{
+		Existing:    existingDs,
+		Error:       loadError,
+		UpdateError: updateError,
+		Desired:     desiredDaemonSet,
+		Compare:     compare.Object,
+		Merge:       merge.ObjectForUpdate,
+	})
+	return ret
+}
